% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmark.R
\docType{data}
\name{Benchmark}
\alias{Benchmark}
\title{Benchmark Class}
\format{\code{\link{R6Class}} object}
\usage{
Benchmark
}
\description{
Benchmark Class
}
\seealso{
Other Benchmark: \code{\link{BenchmarkExpensive}}
}
\keyword{datasets}
