% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkExpensive.R
\docType{data}
\name{BenchmarkExpensive}
\alias{BenchmarkExpensive}
\title{BenchmarkExpensive Class}
\format{\code{\link{R6Class}} object}
\usage{
BenchmarkExpensive
}
\description{
BenchmarkExpensive Class
}
\seealso{
Other Benchmark: \code{\link{Benchmark}}
}
\keyword{datasets}
