% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchMultiVis.R
\docType{data}
\name{BenchMultiVis}
\alias{BenchMultiVis}
\title{Visualize and Evaluate multiple Benchmarks}
\format{\code{\link{R6Class}} object}
\usage{
BenchMultiVis
}
\arguments{
\item{res.list}{[\code{list}]\cr
List of BenchmarkResult Objects}

\item{benchmarks}{[\code{list}]\cr
List of Benchmarks.
Will be matched with res.list entries.}
}
\description{
Visualize and Evaluate multiple Benchmarks
}
\keyword{datasets}
