% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchReplVis.R
\name{BenchReplVis}
\alias{BenchReplVis}
\title{Visualization for a single repeated Benchmar}
\usage{
BenchReplVis(res.list, benchmark)
}
\arguments{
\item{res.list}{[\code{list}]\cr
Lift of BenchResult Objects}

\item{benchmark}{[\code{Benchmark}]\cr
The Benchmark object for the benchmarks in the \code{res.list}.}
}
\value{
list of ggplot objects
}
\description{
Visualization for a single repeated Benchmar
}
