% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOmlBotBenchmark.R
\name{generateOmlBotBenchmark}
\alias{generateOmlBotBenchmark}
\title{Generates a simple Benchmark}
\usage{
generateOmlBotBenchmark(task.id, learner.name)
}
\arguments{
\item{task.id}{[\code{numeric(1)}]\cr
OML Task ID (e.g. 3)}

\item{learner.name}{[\code{character(1)}]\cr
mlr learner name (e.g. classif.svm)}
}
\value{
\code{\link{Benchmark}}
}
\description{
Generates a simple Benchmark from simple inexpensive single objective functions.
}
