% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSimpleBenchmark.R
\name{generateSimpleBenchmark}
\alias{generateSimpleBenchmark}
\title{Generates a simple Benchmark}
\usage{
generateSimpleBenchmark(smoof.fun)
}
\arguments{
\item{smoof.fun}{[\code{smoof_function}]\cr
Single objective smoof function.}
}
\value{
\code{\link{Benchmark}}
}
\description{
Generates a simple Benchmark from simple inexpensive single objective functions.
}
