% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSimpleBenchmarkPreEvaluated.R
\name{generateSimpleBenchmarkPreEvaluated}
\alias{generateSimpleBenchmarkPreEvaluated}
\title{Generates a simple Benchmark}
\usage{
generateSimpleBenchmarkPreEvaluated(smoof.fun, design)
}
\arguments{
\item{smoof.fun}{[\code{smoof_function}]\cr
Single objective smoof function.}

\item{design}{[\code{data.frame}]\cr
The preevaluated design of the function.
The results have to be in the column "y"}
}
\value{
\code{\link{Benchmark}}
}
\description{
Generates a simple Benchmark from simple inexpensive single objective functions.
}
