% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminationCriterion.R
\docType{data}
\name{TerminationCriterion}
\alias{TerminationCriterion}
\title{TerminationCriterion Class}
\format{\code{\link{R6Class}} object}
\usage{
TerminationCriterion
}
\description{
TerminationCriterion Class
}
\seealso{
Other TerminationCriterion: \code{\link{TerminationEvals}},
  \code{\link{TerminationIteration}},
  \code{\link{TerminationValue}}
}
\keyword{datasets}
