% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_criterions.R
\docType{data}
\name{TerminationEvals}
\alias{TerminationEvals}
\title{TerminationEvals Class}
\format{\code{\link{R6Class}} object}
\usage{
TerminationEvals
}
\description{
TerminationEvals Class
}
\seealso{
Other TerminationCriterion: \code{\link{TerminationCriterion}},
  \code{\link{TerminationIteration}},
  \code{\link{TerminationValue}}
}
\concept{TerminationCriterion}
\keyword{datasets}
