% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_criterions.R
\docType{data}
\name{TerminationIteration}
\alias{TerminationIteration}
\title{TerminationIteration Class}
\format{\code{\link{R6Class}} object}
\usage{
TerminationIteration
}
\description{
TerminationIteration Class
}
\seealso{
Other TerminationCriterion: \code{\link{TerminationCriterion}},
  \code{\link{TerminationEvals}},
  \code{\link{TerminationValue}}
}
\concept{TerminationCriterion}
\keyword{datasets}
