% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_criterions.R
\docType{data}
\name{TerminationValue}
\alias{TerminationValue}
\title{TerminationValue Class}
\format{\code{\link{R6Class}} object}
\usage{
TerminationValue
}
\description{
TerminationValue Class
}
\seealso{
Other TerminationCriterion: \code{\link{TerminationCriterion}},
  \code{\link{TerminationEvals}},
  \code{\link{TerminationIteration}}
}
\concept{TerminationCriterion}
\keyword{datasets}
