% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchMultiVis.R
\name{BenchMultiVis}
\alias{BenchMultiVis}
\title{Visualize and Evaluate multiple Benchmarks}
\usage{
BenchMultiVis(res.list, benchmarks)
}
\arguments{
\item{res.list}{[\code{list}]\cr
List of BenchmarkResult Objects}

\item{benchmarks}{[\code{list}]\cr
List of Benchmarks.
Will be matched with res.list entries.}
}
\value{
list of ggplots and tables
}
\description{
Visualize and Evaluate multiple Benchmarks
}
